Add dev dependencies
So! first serious commit :)

The first choice for this project is to just use the default http server
from node. I don't know which http server will we require, so the
default one is good enough. That also shows that I know how to use the
node raw http server. If I do my job correctly, the app / business code should be
able to swap external dependencies without much hassle.

I also use typescript@beta (5.0.0), because of its development performance
improvements. Next week it's going to get out of beta, so it's not
really a big issue to use this prerelease version for now.