Manually add dev boilerplate
I really enjoy having simple dev setups. The esbuild dependency does all
for me, it has a really simple dev interface.
Also, node 18 has a built in file watcher, so we have built in live
reloading! We can trim the `nodemon` dependency.

With these two tools we have our basic development toolkit to continue.
We just need to setup tests with jest and we'll be done!