Define repository errors
I'm a believer that errors are part of the domain. Golang taught me that
errors as values result into more explicit and, as a result, simpler
control flow. That's why I don't use exceptions in the core logic.
Typescript is powerful enought to allow for discriminated unions, aka
"sum types". That allows me to define the diferent errors that can
happen in our REST API. That's also the reason why the "assertNever"
helper I defined is so useful, it allows for exhaustive pattern
matching.