Refactor tests to enforce retrieval of latest reviews
So, the requirements are ambiguous about what to present on the reviews.
What I'm pretty sure is that a business possibly can have a huge amount
of reviews, so presenting all of them to the business details endpoint
would be totally wasteful and unnecessary.

I decided to just present the latest reviews, and add a creation date to
allow for sorting and filtering by latest date. The creation date is
always useful data, so if this decision doesn't really fit with the
business requirements we could at least keep that part of the logic.